feat: ask for permission before deleting

+++ assets/scripts/free_space.sh
@@ -2,7 +2,9 @@
 
 for dir in ~/Documents/{GitHub,Sources}/{*,*/*}/{.flatpak-builder,builddir,build,dist,node_modules,target}; do
   if [[ -d "$dir" ]]; then
-    echo "Deleting $dir"
-    rm -rf "$dir"
+    read -p "Delete $dir? (y/N) "
+    if [[ $REPLY =~ ^[Yy] ]]; then
+      rm -rf "$dir"
+    fi
   fi
 done
